+++ crates/aiken/src/cmd/blueprint/address.rs
@@ -1,4 +1,3 @@
-use aiken_lang::ast::Tracing;
 use aiken_project::watch::with_project;
 use std::path::PathBuf;
 
@@ -20,10 +19,6 @@ pub struct Args {
     #[clap(long)]
     delegated_to: Option<String>,
 
-    /// Force the project to be rebuilt, otherwise relies on existing artifacts (i.e. plutus.json)
-    #[clap(long)]
-    rebuild: bool,
-
     /// Output the address for mainnet (this command defaults to testnet)
     #[clap(long)]
     mainnet: bool,
@@ -35,15 +30,10 @@ pub fn exec(
         module,
         validator,
         delegated_to,
-        rebuild,
         mainnet,
     }: Args,
 ) -> miette::Result<()> {
     with_project(directory.as_deref(), false, |p| {
-        if rebuild {
-            p.build(false, Tracing::silent())?;
-        }
-
         let title = module.as_ref().map(|m| {
             format!(
                 "{m}{}",